`Bugfix`: Show exercise title (#126)
